requirements.txt
discord.py==2.6.4
aiohttp==3.13.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.9
